OLWIDGET_DEFAULT_OPTIONS in settings.
If we provide the ability to set global, default options for widget
construction then it's possible to build reusable apps on top of
olwidget types.  For instance, if you write an app that returns an
InfoMap widget then you can be assured it will be styled in a roughly
similar fashion to your other InfoMaps.